Add implementation for insert or update if exists in BCQ data
 - refactor sql queries for BCQ data
 - add exists method for checking if data exists
 - include changing datatype of BCQ to float to support 9 digits after decimal point
 - include adding of unique key for selling_mtn, buying_participant, and end_time

+++ metering-app/src/main/java/com/pemc/crss/metering/dto/BCQData.java
@@ -15,7 +15,7 @@ public class BCQData {
     private String referenceMTN;
     private Date startTime;
     private Date endTime;
-    private double bcq;
+    private float bcq;
 
     @Override
     public String toString() {
