+++ metering-app/src/main/java/com/pemc/crss/metering/dto/BCQDataInfo.java
@@ -12,6 +12,6 @@ public class BCQDataInfo {
     private String referenceMTN;
     private Date startTime;
     private Date endTime;
-    private double bcq;
+    private float bcq;
 
 }
